Add image traverse script
Travel from one image to another and observe the change in labels and
features.

Select correctly-classified images and upgrade the accuracy of the distance vector

The selected images for traverse must be correctly classified by the
neural network classifier.
The distance vector must be double, otherwise the truncation error for
uint8 is significant