Add codes and stats for CleverHans demo
Adversarial training section is skipped because we're more interested in the impact from those illusive samples regarding the "intrinsic" robustness of classifiers trained with clean samples. 
The stats are precomputed offline and stored in a .pkl file, which includes the number of times a training samples has been classified correctly during training. 
Note: the scripts are still at a primitive stage which requires manual configuration for different set-ups.